+++ opensearchapi/api_index.go
@@ -30,6 +30,14 @@ func (c Client) Index(ctx context.Context, req IndexReq) (*IndexResp, error) {
 		return &data, err
 	}
 
+	if c.returnQueryErrors && data.Shards.Failed > 0 {
+		return &data, &ShardFailureError{
+			Operation:    OperationIndex,
+			FailedShards: data.Shards.Failed,
+			TotalShards:  data.Shards.Total,
+		}
+	}
+
 	return &data, nil
 }