+++ opensearchapi/api_search-template.go
@@ -26,6 +26,14 @@ func (c Client) SearchTemplate(ctx context.Context, req SearchTemplateReq) (*Sea
 		return &data, err
 	}
 
+	if c.returnQueryErrors && data.Shards.Failed > 0 {
+		return &data, &PartialSearchError{
+			FailedShards: data.Shards.Failed,
+			TotalShards:  data.Shards.Total,
+			Failures:     data.Shards.Failures,
+		}
+	}
+
 	return &data, nil
 }